+++ display/fontgen/out/font_8_cyrillic.c
@@ -584,6 +584,42 @@ static const uint8_t bmp_8_cyrillic_0451[] = {
 0x60, 
 };
 
+// U+00B0 '°'
+static const uint8_t bmp_8_cyrillic_00B0[] = {
+0x40, 
+0x40, 
+};
+
+// U+00B5 'µ'
+static const uint8_t bmp_8_cyrillic_00B5[] = {
+0x90, 
+0x90, 
+0x90, 
+0xF0, 
+0x80, 
+};
+
+// U+00B2 '²'
+static const uint8_t bmp_8_cyrillic_00B2[] = {
+0xC0, 
+0x40, 
+0xE0, 
+};
+
+// U+00B3 '³'
+static const uint8_t bmp_8_cyrillic_00B3[] = {
+0xC0, 
+0x40, 
+0xE0, 
+};
+
+// U+2082 '₂'
+static const uint8_t bmp_8_cyrillic_2082[] = {
+0xC0, 
+0x40, 
+0xE0, 
+};
+
 static const Glyph glyphs_8_cyrillic[] = {
   { 0x0401, 4, 7, 0, bmp_8_cyrillic_0401 }, // 'Ё'
   { 0x0410, 5, 5, 2, bmp_8_cyrillic_0410 }, // 'А'
@@ -651,8 +687,13 @@ static const Glyph glyphs_8_cyrillic[] = {
   { 0x044E, 6, 4, 3, bmp_8_cyrillic_044E }, // 'ю'
   { 0x044F, 4, 4, 3, bmp_8_cyrillic_044F }, // 'я'
   { 0x0451, 4, 5, 2, bmp_8_cyrillic_0451 }, // 'ё'
+  { 0x00B0, 3, 2, 2, bmp_8_cyrillic_00B0 }, // '°'
+  { 0x00B5, 4, 5, 3, bmp_8_cyrillic_00B5 }, // 'µ'
+  { 0x00B2, 3, 3, 1, bmp_8_cyrillic_00B2 }, // '²'
+  { 0x00B3, 3, 3, 1, bmp_8_cyrillic_00B3 }, // '³'
+  { 0x2082, 3, 3, 5, bmp_8_cyrillic_2082 }, // '₂'
 };
 
 const Font font_8_cyrillic = {
-  66, 8, glyphs_8_cyrillic
+  71, 8, glyphs_8_cyrillic
 };